pyproject.toml
[tool.poetry]
name = "acesdatamanager"
version = "0.1.0"
description = ""
authors = ["kapsali29 <kapsali29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymgclient = "^1.3.1"
psycopg2-binary = "^2.9.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
confluent-kafka = "^2.3.0"
neo4j = "^5.17.0"
minio = "^7.2.5"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

